rework mdx components
