+++ apps/badass/src/components/mdx/components/index.ts
@@ -1,0 +1,14 @@
+export {
+  default as BodyBlockquote,
+  type BodyBlockquoteProps,
+} from './body-blockquote'
+
+export {default as BodyImage, type BodyImageProps} from './body-image'
+
+export {
+  default as RelatedTeamMembers,
+  TeamMemberCard,
+  type TeamMemberCardProps,
+} from './team-members'
+
+export {default as TweetEmbed, type TweetEmbedProps} from './tweet-embed'
