+++ apps/badass/src/components/mdx/index.tsx
@@ -1,0 +1,34 @@
+import {
+  BodyBlockquote,
+  BodyBlockquoteProps,
+  BodyImage,
+  BodyImageProps,
+  RelatedTeamMembers,
+  TeamMemberCard,
+  TeamMemberCardProps,
+  TweetEmbed,
+  TweetEmbedProps,
+} from './components'
+
+const mdxComponents = {
+  BodyBlockquote: ({
+    children,
+    color,
+  }: React.PropsWithChildren<BodyBlockquoteProps>) => {
+    return <BodyBlockquote color={color}>{children}</BodyBlockquote>
+  },
+  BodyImage: ({src, width, height, alt}: BodyImageProps) => {
+    return <BodyImage src={src} width={width} height={height} alt={alt} />
+  },
+  TeamMemberCard: ({imageUrl, name, title}: TeamMemberCardProps) => {
+    return <TeamMemberCard imageUrl={imageUrl} name={name} title={title} />
+  },
+  RelatedTeamMembers: ({children}: React.PropsWithChildren) => {
+    return <RelatedTeamMembers>{children}</RelatedTeamMembers>
+  },
+  TweetEmbed: ({tweetId}: TweetEmbedProps) => {
+    return <TweetEmbed tweetId={tweetId} />
+  },
+}
+
+export default mdxComponents
