+++ apps/badass/src/templates/case-study-template.tsx
@@ -1,19 +1,19 @@
 import React from 'react'
 import {PortableText, toPlainText} from '@portabletext/react'
 import {LinkedIn, Twitter} from '@skillrecordings/react'
-import {CalendarIcon} from '@heroicons/react/outline'
-import {SanityDocument} from '@sanity/client'
+// import {CalendarIcon} from '@heroicons/react/outline'
+// import {SanityDocument} from '@sanity/client'
 import {useRouter} from 'next/router'
-import {format} from 'date-fns'
+// import {format} from 'date-fns'
 import JoelHooksHeadshotImage from '../../public/joel-hooks.jpg'
-import PortableTextComponents from 'components/portable-text'
+// import PortableTextComponents from 'components/portable-text'
 import Layout from 'components/layout'
 import Image from 'next/image'
 import {SmallCallToActionForm} from '../components/call-to-action-form'
 import {genericCallToActionContent} from '../components/landing-content'
 import {type MDXRemoteSerializeResult} from 'next-mdx-remote'
 import MDX from '@skillrecordings/skill-lesson/markdown/mdx'
-import mdxComponents from 'components/mdx/components'
+import mdxComponents from 'components/mdx'
 
 import Balancer from 'react-wrap-balancer'
 import type {CaseStudy} from 'lib/case-studies'
